slices moved
