+++ src/components/burger-constructor-element/burger-constructor-element.tsx
@@ -6,7 +6,7 @@ import {
   removeBurgerItem,
   moveBurgerItemUp,
   moveBurgerItemDown
-} from '../../services/slices/constructorSlice';
+} from '../../services/slices/constructor/constructorSlice';
 
 export const BurgerConstructorElement: FC<BurgerConstructorElementProps> = memo(
   ({ ingredient, index, totalItems }) => {
