+++ src/components/order-card/order-card.tsx
@@ -6,7 +6,7 @@ import { useSelector } from '../../services/store';
 import { OrderCardProps } from './type';
 import { TIngredient } from '@utils-types';
 import { OrderCardUI } from '../ui/order-card';
-import { selectIngredients } from '../../services/slices/ingredientsSlice';
+import { selectIngredients } from '../../services/slices/ingredients/ingredientsSlice';
 
 const maxIngredients = 6;
