+++ src/pages/profile/profile.tsx
@@ -1,7 +1,7 @@
 import { ProfileUI } from '@ui-pages';
 import { FC, SyntheticEvent, useEffect, useState } from 'react';
 // import { useDispatch, useSelector } from 'react-redux';
-import { updateUser } from '../../services/slices/authSlice';
+import { updateUser } from '../../services/slices/auth/authSlice';
 import { RootState, useDispatch, useSelector } from '../../services/store';
 
 export const Profile: FC = () => {
