+++ src/services/rootReducer.ts
@@ -1,10 +1,10 @@
 import { combineReducers } from '@reduxjs/toolkit';
-import ingredientsReducer from './slices/ingredientsSlice';
-import constructorReducer from './slices/constructorSlice';
-import feedReducer from './slices/feedSlice';
-import authReducer from './slices/authSlice';
-import { orderReducer } from './slices/orderSlice';
-import userOrdersReducer from './slices/userOrdersSlice';
+import ingredientsReducer from './slices/ingredients/ingredientsSlice';
+import constructorReducer from './slices/constructor/constructorSlice';
+import feedReducer from './slices/feed/feedSlice';
+import authReducer from './slices/auth/authSlice';
+import { orderReducer } from './slices/order/orderSlice';
+import userOrdersReducer from './slices/userOrders/userOrdersSlice';
 
 const rootReducer = combineReducers({
   ingredients: ingredientsReducer,
