+++ src/services/slices/auth/authSlice.ts
@@ -9,7 +9,7 @@ import {
   updateUserApi
 } from '@api';
 import { TUser } from '@utils-types';
-import { deleteCookie, setCookie } from '../../utils/cookie';
+import { deleteCookie, setCookie } from '../../../utils/cookie';
 
 interface AuthState {
   user: TUser | null;
