+++ src/services/slices/ingredients/ingredientsSlice.ts
@@ -1,6 +1,6 @@
 import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
-import { TIngredient } from '../../utils/types';
-import { getIngredientsApi } from '../../utils/burger-api';
+import { TIngredient } from '../../../utils/types';
+import { getIngredientsApi } from '../../../utils/burger-api';
 
 export const fetchIngredients = createAsyncThunk(
   'ingredients/fetchIngredients',
